requirements.txt
BeautifulSoup
Cython==0.20.1
Pillow==2.5.3
-e git+https://github.com/mitocw/latex2dnd.git@master#egg=latex2dnd-master
lxml==3.3.3
path.py==5.1
pep8==1.5.7
-e git+https://github.com/tiarno/plastex@master#egg=plasTeX-master
wsgiref==0.1.2
pyyaml==3.10
